Split `PurchasesTests` step 6: extracted `PurchasesSyncPurchasesTests`
