+++ Tests/UnitTests/Purchasing/Purchases/PurchasesTests.swift
@@ -330,163 +330,6 @@ class PurchasesTests: BasePurchasesTests {
         expect(self.notificationCenter.observers.count).to(equal(2))
     }
 
-    func testSyncPurchasesPostsTheReceipt() {
-        setupPurchases()
-        purchases.syncPurchases(completion: nil)
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-    }
-
-    func testSyncPurchasesPostsTheReceiptIfAutoSyncPurchasesSettingIsOff() throws {
-        systemInfo = try MockSystemInfo(platformInfo: nil,
-                                        finishTransactions: false,
-                                        dangerousSettings: DangerousSettings(autoSyncPurchases: false))
-        self.initializePurchasesInstance(appUserId: nil)
-
-        purchases.syncPurchases(completion: nil)
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-    }
-
-    func testSyncPurchasesDoesntPostIfReceiptEmptyAndCustomerInfoLoaded() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:],
-                "original_application_version": "1.0",
-                "original_purchase_date": "2018-10-26T23:17:53Z"
-            ]])
-
-        let object = try info.asData()
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-
-        mockTransactionsManager.stubbedCustomerHasTransactionsCompletionParameter = false
-
-        setupPurchases()
-        purchases.syncPurchases(completion: nil)
-
-        expect(self.backend.postReceiptDataCalled) == false
-    }
-
-    func testSyncPurchasesPostsIfReceiptEmptyAndCustomerInfoNotLoaded() {
-        mockTransactionsManager.stubbedCustomerHasTransactionsCompletionParameter = false
-
-        setupPurchases()
-        purchases.syncPurchases(completion: nil)
-
-        expect(self.backend.postReceiptDataCalled) == true
-    }
-
-    func testSyncPurchasesPostsIfReceiptHasTransactionsAndCustomerInfoLoaded() throws {
-        let info = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:],
-                "original_application_version": "1.0",
-                "original_purchase_date": "2018-10-26T23:17:53Z"
-            ]])
-
-        let object = try info.asData()
-        self.deviceCache.cachedCustomerInfo[identityManager.currentAppUserID] = object
-
-        mockTransactionsManager.stubbedCustomerHasTransactionsCompletionParameter = true
-
-        setupPurchases()
-        purchases.syncPurchases(completion: nil)
-
-        expect(self.backend.postReceiptDataCalled) == true
-    }
-
-    func testSyncPurchasesPostsIfReceiptHasTransactionsAndCustomerInfoNotLoaded() {
-        mockTransactionsManager.stubbedCustomerHasTransactionsCompletionParameter = true
-
-        setupPurchases()
-        purchases.syncPurchases(completion: nil)
-
-        expect(self.backend.postReceiptDataCalled) == true
-    }
-
-    func testSyncPurchasesDoesntRefreshTheReceiptIfNotEmpty() {
-        setupPurchases()
-        self.receiptFetcher.shouldReturnReceipt = true
-        purchases.syncPurchases(completion: nil)
-
-        expect(self.receiptFetcher.receiptDataTimesCalled) == 1
-        expect(self.requestFetcher.refreshReceiptCalled) == false
-    }
-
-    func testSyncPurchasesDoesntRefreshTheReceiptIfEmpty() {
-        setupPurchases()
-        self.receiptFetcher.shouldReturnReceipt = false
-        purchases.syncPurchases(completion: nil)
-
-        expect(self.receiptFetcher.receiptDataTimesCalled) == 1
-        expect(self.requestFetcher.refreshReceiptCalled) == false
-    }
-
-    func testSyncPurchasesPassesIsRestoreAsAllowSharingAppStoreAccount() {
-        setupPurchases()
-
-        var deprecated = purchases.deprecated
-        deprecated.allowSharingAppStoreAccount = false
-        purchases.syncPurchases(completion: nil)
-        expect(self.backend.postedIsRestore!) == false
-
-        deprecated.allowSharingAppStoreAccount = true
-        purchases.syncPurchases(completion: nil)
-        expect(self.backend.postedIsRestore!) == true
-    }
-
-    func testSyncPurchasesSetsIsRestoreForAnon() {
-        setupAnonPurchases()
-
-        var deprecated = purchases.deprecated
-        deprecated.allowSharingAppStoreAccount = false
-        purchases.syncPurchases(completion: nil)
-        expect(self.backend.postedIsRestore!) == false
-
-        deprecated.allowSharingAppStoreAccount = true
-        purchases.syncPurchases(completion: nil)
-        expect(self.backend.postedIsRestore!) == true
-    }
-
-    func testSyncPurchasesCallsSuccessDelegateMethod() throws {
-        setupPurchases()
-
-        let customerInfo = try CustomerInfo(data: Self.emptyCustomerInfoData)
-        self.backend.postReceiptResult = .success(customerInfo)
-
-        var receivedCustomerInfo: CustomerInfo?
-
-        purchases!.syncPurchases { (info, _) in
-            receivedCustomerInfo = info
-        }
-
-        expect(receivedCustomerInfo).toEventually(be(customerInfo))
-    }
-
-    func testSyncPurchasesPassesErrorOnFailure() {
-        setupPurchases()
-
-        let error: BackendError = .missingAppUserID()
-
-        self.backend.postReceiptResult = .failure(error)
-        self.purchasesDelegate.customerInfo = nil
-
-        var receivedError: Error?
-
-        purchases!.syncPurchases { (_, newError) in
-            receivedError = newError
-        }
-
-        expect(receivedError).toEventuallyNot(beNil())
-        expect(receivedError).to(matchError(error.asPurchasesError))
-    }
-
     func testCallsShouldAddPromoPaymentDelegateMethod() {
         setupPurchases()
         let product = StoreProduct(sk1Product: MockSK1Product(mockProductIdentifier: "mock_product"))
@@ -903,18 +746,6 @@ class PurchasesTests: BasePurchasesTests {
         Purchases.proxyURL = nil
 
         expect(SystemInfo.serverHostURL) == defaultHostURL
-    }
-
-    @available(iOS 14.0, macOS 14.0, tvOS 14.0, watchOS 7.0, *)
-    func testSyncsPurchasesIfEntitlementsRevokedForProductIDs() throws {
-        try AvailabilityChecks.iOS14APIAvailableOrSkipTest()
-
-        setupPurchases()
-        guard purchases != nil else { fatalError() }
-        expect(self.backend.postReceiptDataCalled).to(beFalse())
-        (purchasesOrchestrator as StoreKitWrapperDelegate)
-            .storeKitWrapper(storeKitWrapper, didRevokeEntitlementsForProductIdentifiers: ["a", "b"])
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
     }
 
     @available(*, deprecated) // Ignore deprecation warnings
